build: Add subcommands & Fix PuTTY install

+++ start_robot.ps1
@@ -1,12 +1,21 @@
+param (
+    [Alias("r")][switch]$RunOnly,
+    [Alias("p")][switch]$PushOnly
+)
+if ($RunOnly -and $PushOnly) {
+    Write-Host "ERROR: You cannot use both -RunOnly and -PushOnly at the same time." -ForegroundColor Red
+    exit 1
+}
+
+
 function Check-Command($cmdname)
 {
     return [bool](Get-Command -Name $cmdname -ErrorAction SilentlyContinue)
 }
-
 if (-not (Check-Command -cmdname 'plink'))
 {
      Write-Host "PuTTy not found. Installing..."
-     winget install putty --disable-interactivity
+     winget install PuTTY.PuTTY --disable-interactivity
 }
 
 
@@ -28,20 +37,26 @@ if (-not (Test-Path -Path "./logs"))
 
 
 $password = "maker"
+if (-not $RunOnly)
+{
+    # PUSH SOURCE CODE TO EV3DEV
+    Write-Host "Pushing code to EV3DEV..."
+    $ev3dev_launch_script = "rm -r /home/robot/SRA_G4_P3/*"
+    plink -batch -ssh robot@ev3dev -pw $password $ev3dev_launch_script
+    pscp -batch -r -pw $password ./src robot@ev3dev:/home/robot/SRA_G4_P3
+    $ev3dev_launch_script = "chmod +rx /home/robot/SRA_G4_P3/src/*.py"
+    plink -batch -ssh robot@ev3dev -pw $password $ev3dev_launch_script
+}
+if (-not $PushOnly)
+{
+    # LAUNCH SCRIPT WITHIN EV3DEV
+    Write-Host "Launching EV3DEV program..."
+    $ev3dev_launch_script = "python3 /home/robot/SRA_G4_P3/src/main.py"
+    plink -batch -ssh robot@ev3dev -pw $password $ev3dev_launch_script
 
-# PUSH SOURCE CODE TO EV3DEV
-Write-Host "Pushing code to EV3DEV..."
-$ev3dev_launch_script = "rm -r /home/robot/SRA_G4_P3/*"
-plink -batch -ssh robot@ev3dev -pw $password $ev3dev_launch_script
-pscp -batch -r -pw $password ./src robot@ev3dev:/home/robot/SRA_G4_P3
+    # GET LOGS
+    Write-Host "Downloading logs..."
+    pscp -batch -pw $password robot@ev3dev:/home/robot/SRA_G4_P3/*.log ./logs/
+}
 
-
-# LAUNCH SCRIPT WITHIN EV3DEV
-Write-Host "Launching EV3DEV program..."
-$ev3dev_launch_script = "chmod +rx /home/robot/SRA_G4_P3/src/*.py && python3 /home/robot/SRA_G4_P3/src/main.py"
-plink -batch -ssh robot@ev3dev -pw $password $ev3dev_launch_script
-
-
-# GET LOGS
-Write-Host "Downloading logs..."
-pscp -batch -pw $password robot@ev3dev:/home/robot/SRA_G4_P3/*.log ./logs/
+Write-Host "Done!"
